refactor: pkg data-analysis
