+++ lib/data-analysis/cosine-similarity.js
@@ -1,5 +1,3 @@
-/** @module lib/data-analysis */
-
 /**
  * Function to calculate the dot product of two vectors
  *
