+++ lib/data-analysis/word-frequency.js
