+++ lib/db/migrations.js
@@ -1,5 +1,5 @@
 import { ObjectID } from "bson";
-import { getInsights } from "../insights";
+import { getInsights } from "../data-analysis/word-frequency";
 import { getPostsCollection } from "../mongodb";
 import { decryptPost } from "../security/transformations";
 import { hasCommentedOnPost } from "./reads";
