+++ lib/db/posts/writes.js
@@ -1,5 +1,5 @@
 import { ObjectID } from "bson";
-import { getInsights } from "../../insights";
+import { getInsights } from "../../data-analysis/word-frequency";
 import {
   getCommentsCollection,
   getPostsCollection,
